fix(materials): only contains mat information now
