+++ qmt/materials/__init__.py
@@ -1,2 +1,3 @@
 from .materials import Material, Materials, conduction_band_offset, valence_band_offset
 from .mat_builder import build_materials, make_materials_library
+from .mat_data import MatData, MatPart
